Moved code of online provider form eventactor to Provider.scala in event-mgmt.
